+++ forge/src/main/java/com/li64/tide/TideForgeEvents.java
@@ -108,6 +108,7 @@ public class TideForgeEvents {
         @SubscribeEvent
         public static void onLootTableLoad(LootTableLoadEvent event) {
             if (event.getName().toString().equals("minecraft:gameplay/fishing")) {
+                if (Tide.CONFIG.general.crateWeight <= 0) return;
                 // Add crate rolls
                 LootPool pool = ((LootTableAccessor) event.getTable()).tide$getPool(0);
                 ((LootTableAccessor) event.getTable()).tide$getPool(0).entries = new ImmutableList.Builder<LootPoolEntryContainer>()